requirements.txt
requests>=2.28.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# LangGraph Agent 依赖
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
